Update options page to look nicer
A whole lot of cleanup in this. Pulled out all the of the options page html and
put it into a seperate file and also completely re-wrote the majority of it. Matches
the CJ plugin stylistically.

Also did a lot of general cleanup on the main file (which inexplicably had no
indentation...).

Redmine-Ticket: 1975