Try to create cache directory if it does not exist
If the cache directory does not exist and does not have the proper
permissions, then users get an error to that effect.  Here we test for
the existence of the cache directory and then try to create it if
needed.

This itself can fail, if either the plugin script is lacking proper
permissions, of if the directory containing the cache directory does
not have the right permissions.  To help with this, these changes also
try to be a little more aware of possible permission failings,
checking the results of file writes and such.

Redmine-Ticket: 2017